gh-113317: Move FormatCounterFormatter into libclinic
